requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
umap-learn>=0.5.4
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
datasets>=2.14.0
soundfile>=0.12.0
torchcodec
librosa>=0.10.0
pydub>=0.25.0